feat: added error handling

+++ demo/src/main/java/com/example/demo/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.example.demo.student;
 
+import com.example.demo.student.exeptions.BadRequestException;
+import com.example.demo.student.exeptions.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,11 +17,19 @@ public class StudentService {
         return studentRepository.findAll();
     }
 
-    public void addStudent(Student student) {
+    public void addStudent(Student student) throws BadRequestException {
+        Boolean existEmail = studentRepository.selectExistsEmail(student.getEmail());
+
+        if(existEmail) {
+            throw new BadRequestException("Student already exists");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException("Student not found");
+        }
         studentRepository.deleteById(studentId);
     };
